Move node state transition captor helper action to ModelTestActions
